+++ types/x-template.d.ts
@@ -1,3 +1,3 @@
 export const render: (container: HTMLElement, rawResult: unknown) => void;
-export const html: (strings: string[], ...values: unknown[]) => unknown;
+export const html: (strings: TemplateStringsArray, ...values: unknown[]) => unknown;
 //# sourceMappingURL=x-template.d.ts.map