+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Changed
+- Fix `noImplicitAny` TypeScript errors in `x-template.js` (#357).
 - Fix `noImplicitAny` TypeScript errors in `x-parser.js` (#357).
 - Re-enable strict type errors for `strictNullChecks`,
   `useUnknownInCatchVariables`, and `strictFunctionTypes` (#357).